Combining collectd data with metadata.
Metadata is retrieved from either of the cache or the remote API
and dropped into a Packet struct.  This struct is copied for each
collectd packet in an incoming collectd message, and the collectd
data appended to it.

When a collectd message has >1 value, the collectd Packet is
duplicated once for each value.