Added, then removed QA test harness.
Initially, the plan was to create an HTTP server that exposed a
port through which QA could POST messages into the system, without
relying on Kafka.

On the producer side, the utility would be configured to point to a
callback URL, where it would POST the decorated string.

This approach ultimately required that `if QA.enabled then` statements
be littered throughout the code, and it bypassed a lot of the logic
pertaining to establishing and managing connections.  So I threw it
out.

Now, the QA branch of the configuration exposes two options -
enabled, and server_count.  enabled turns on server randomization,
and server_count sets the number of randomized servers that will be
associated with each incoming metric.  This is useful in the event
that the server you are using to generate test metrics is not
represented with a decorated string.  Metrics from that server can
be decorated at random with an arbitrary match to other servers.